arch: add support for mcf5441x Colfdire
The m68k family ony has one Coldfire variant, namely the 5208. Add the
support for the MCF5441x CPU family in the configuration file.
This covers MCF54410, MCF54415, MCF54416, MCF54417 and MCF54418.

Commit [1] removed the GCC ARCH configuration which was used to
determine if the architecture is "m68k" or "cf". I have re-introduced
this configuration here, only for the MCF5441x case. It has been tested
as an external toolchain too.

[1]: eda11417be m68k: remove BR2_GCC_TARGET_ARCH

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>