board: add the qemu board support for mcf54418
The Coldfire MCF54418 needs a board to give the proper linux defconfig
to buildroot. Add it based on a simplified Linux stmark2_defconfig file.
While at it, add a the needed config to run qemu, using a copy of the
qemu_m68k_mcf5208_defconfig as there is not so much difference to expect
except the CPU value.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>