package: pcre2: fix m68k ColdFire build error
The assembler was failing with the error:

/tmp/ccefZbef.s: Assembler messages:
/tmp/ccefZbef.s: Fatal error: Tried to convert PC relative branch to absolute jump

This issue occurs when compiling for the m68k ColdFire architecture. The
problem stems from the use of 16-bit offsets in "switch" tables, which
can lead to PC-relative branch conversion errors.

To resolve this, the -mlong-jump-table-offsets flag is added to the
CFLAGS for ColdFire targets. This forces the use of 32-bit offsets in
"switch" tables, preventing the assembler from converting PC-relative
branches to absolute jumps.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>