package/uclibc: fix NPTL alignment on m68k
Tested on QEMU using the qemu_m68k_q800_defconfig configuration, with a
switch of  the toolchain to uclibc-ng and enabling rsyslog.

This patch addresses an issue where rsyslog’s pthread_cond_wait was
busy-looping due to futex_time64 repeatedly returning EINVAL. The
correction in NPTL alignment resolves the 100% CPU usage problem.

This patch is fixed upstream [1] so apply it until it is tagged.

[1]: 278ac6b30 ("m68k: fix alignment in NPTL code")

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>